+++ lib/src/sigv4.dart
@@ -35,13 +35,9 @@ class Sigv4 {
     return digest.bytes;
   }
 
-  /// Hex-encodes a SHA256 hash of the given canonical request string, `request`
-  static String hashCanonicalRequest(String request) {
-    return hexEncode(hash(utf8.encode(request)));
-  }
-
-  static String hashPayload(dynamic payload) {
-    return hexEncode(hash(utf8.encode(payload.toString())));
+  /// Hex-encodes a SHA256 hash of the given payload string, `payload`
+  static String hashPayload(String payload) {
+    return hexEncode(hash(utf8.encode(payload)));
   }
 
   /// Encodes the given string to a canonical URI
@@ -72,17 +68,15 @@ class Sigv4 {
 
   /// Builds a canonical header string from the given `headers`
   static String buildCanonicalHeaders(Map<String, dynamic> headers) {
-    final sortedKeys = [];
-    headers.forEach((property, _) {
-      sortedKeys.add(property);
-    });
+    final sortedKeys = <String>[];
+    headers.forEach((property, _) => sortedKeys.add(property));
 
     var canonicalHeaders = '';
     sortedKeys.sort((a, b) => a.toLowerCase().compareTo(b.toLowerCase()));
 
     sortedKeys.forEach((property) {
       canonicalHeaders +=
-          '${property.toLowerCase()}:${headers[property].toString()}\n';
+          '${property.toLowerCase().trim()}:${headers[property].toString().trim()}\n';
     });
 
     return canonicalHeaders;
@@ -101,8 +95,11 @@ class Sigv4 {
 
   /// Builds the complete canonical string to sign
   static String buildStringToSign(
-      String datetime, String credentialScope, String hashedCanonicalRequest) {
-    return '$_aws_sha_256\n$datetime\n$credentialScope\n$hashedCanonicalRequest';
+    String datetime,
+    String credentialScope,
+    String hashedCanonicalRequest,
+  ) {
+    return '$_aws_sha_256\n${datetime}\n$credentialScope\n$hashedCanonicalRequest';
   }
 
   /// Builds the required credential scope
@@ -131,29 +128,43 @@ class Sigv4 {
   }
 
   /// Builds the `Authorization` headers
-  static String buildAuthorizationHeader(String accessKey,
-      String credentialScope, Map<String, dynamic> headers, String signature) {
+  static String buildAuthorizationHeader(
+    String accessKey,
+    String credentialScope,
+    Map<String, dynamic> headers,
+    String signature,
+  ) {
     return _aws_sha_256 +
         ' Credential=' +
         accessKey +
         '/' +
         credentialScope +
-        ', SignedHeaders=' +
+        ',SignedHeaders=' +
         buildCanonicalSignedHeaders(headers) +
-        ', Signature=' +
+        ',Signature=' +
         signature;
   }
 
   /// Builds the key to use for signing
   static List<int> calculateSigningKey(
-      String secretKey, String datetime, String region, String service) {
+    String secretKey,
+    String datetime,
+    String region,
+    String service,
+  ) {
     return sign(
+      sign(
         sign(
-            sign(
-                sign(utf8.encode('$_aws4$secretKey'), datetime.substring(0, 8)),
-                region),
-            service),
-        _aws4_request);
+          sign(
+            utf8.encode('$_aws4$secretKey'),
+            datetime.substring(0, 8),
+          ),
+          region,
+        ),
+        service,
+      ),
+      _aws4_request,
+    );
   }
 
   /// Builds a Hex-encoded signature with the given `signingKey`and `stringToSign`